rag-project/requirements.txt
fastapi
uvicorn
pydantic
streamlit
requests
openai
chromadb
python-dotenv
tiktoken
plotly
streamlit-mermaid